requirements.txt
black
ruff
mypy
tox
pytest
pytest-snapshot
structlog
pytest-structlog
line_profiler
memory_profiler
pandas[performance, plot, computation, excel, xml, postgresql, hdf5]
polars-lts-cpu
statsmodels
japanize-matplotlib
seaborn
fastcluster
statannotations
python-dotenv
scikit-learn
rdkit
biopython
bioinfokit
scanpy
pyopenms
pyteomics
jupyterlab
ipywidgets
ipympl
streamlit
pygwalker
